Canary release
Commands:
    $ git fetch origin
    $ git checkout -B ui-canary -t origin/ui-canary
    $ git merge --strategy=ours origin/main
    $ git diff --binary origin/main | git apply --reverse --index
    $ git commit --amend

Check end state is correct:
    $ git diff ui-canary origin/main
    [no output]
    $ git rev-list --count ui-canary..origin/main
    0

Then upload:
   git cl issue 0
   git cl upload --bypass-hooks -o banned-words~skip

+++ src/trace_processor/perfetto_sql/stdlib/android/dvfs.sql
@@ -68,9 +68,9 @@ CREATE PERFETTO TABLE android_dvfs_counter_stats(
   -- Counter name on which all the other values are aggregated on.
   name STRING,
   -- Max of all counter values for the counter name.
-  max INT,
+  max DOUBLE,
   -- Min of all counter values for the counter name.
-  min INT,
+  min DOUBLE,
   -- Duration between the first and last counter value for the counter name.
   dur INT,
   -- Weighted avergate of all the counter values for the counter name.
